ssl: split SslSocket to its own target
