+++ source/common/ssl/BUILD
@@ -12,14 +12,24 @@ envoy_cc_library(
     name = "connection_lib",
     srcs = ["connection_impl.cc"],
     hdrs = ["connection_impl.h"],
+    deps = [
+        ":context_lib",
+        ":ssl_socket_lib",
+        "//source/common/network:connection_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "ssl_socket_lib",
+    srcs = ["ssl_socket.cc"],
+    hdrs = ["ssl_socket.h"],
     external_deps = ["ssl"],
     deps = [
         ":context_lib",
+        "//include/envoy/network:transport_socket_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:empty_string",
-        "//source/common/common:hex_lib",
-        "//source/common/network:connection_lib",
-        "//source/common/network:utility_lib",
+        "//source/common/common:logger_lib",
     ],
 )
